requirements.txt
fastapi
uvicorn[standard]
pyautogen